Add files via upload
In a game that moves on the number board the rules are as follows:
The number table consists of N rows and M columns; Rows are numbered from 1 to N, from top to bottom; The columns are numbered 1 to M, from left to right. The cell in the uth row intersecting the vth column is called cell (u,v). Cell (u,v) contains a non-negative integer A(uv).
From cell (u,v), the player can move to a cell with the same edge: (u-1,v), (u+1,v), (u,v-1), (u,v+1) or move to another cell with the same value and cannot move to a cell with value equal to 0. Each move costs one unit of time.
Requirement: Given the location of the starting square and the destination square, find the minimum time to travel from the starting square to the destination square according to the rules of the game.
Data:
- The first line contains two positive integers N and M which are the number of rows and columns of the table.
- The second line includes four numbers x, y, z, t describing the starting point in cell (x,y) and the destination in cell (z,t).
- The following N lines, each line contains M non-negative integers describing the number table.
* Each integer is separated by a space. Data ensures there is always a path from starting point to destination.
Result:
- Contains a positive integer that is the smallest number of time units to go from the starting cell to the destination cell that satisfies the requirements.